// src/app/core/models/ms-chat-clarifications/clarification-goods-reject-notifi-model.ts
export interface IClarificationGoodsReject {
  goodId: number;
  rejectionDate: Date;
  observations: string;
  version: number;
  creationUser: string;
  modificationUser: string;
  creationDate: Date;
  modificationDate: Date;
  rejectNotificationId: number;
  clarificationId: number;
  clarificationType: string;
  reason: string;
  answered: string;
  goodResDevId: number;
  sysStsc4qbnbolbp3szu27tgt6lu: number;
  documentClarificationId: number;
}